API-54 test: adiciona testes automatizados

+++ src/main/java/net/centroweg/gerenciamentocompras/modules/cr/service/crservice/DeleteCr.java
@@ -1,8 +1,10 @@
 package net.centroweg.gerenciamentocompras.modules.cr.service.crservice;
 
 import lombok.RequiredArgsConstructor;
+import net.centroweg.gerenciamentocompras.modules.cr.domain.exception.CrNotFoundException;
 import net.centroweg.gerenciamentocompras.modules.cr.infrastructure.persistence.CrRepository;
 import net.centroweg.gerenciamentocompras.modules.cr.service.mapper.CrMapper;
+import net.centroweg.gerenciamentocompras.shared.MessageDTO;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -11,7 +13,11 @@ public class DeleteCr {
     private final CrRepository crRepository;
     private final CrMapper crMapper;
 
-    public void delete(long id){
+    public MessageDTO delete(long id){
+        if (!crRepository.existsById(id)) {
+            throw new CrNotFoundException(id);
+        }
         crRepository.deleteById(id);
+        return new MessageDTO("CR Deletado com sucesso!");
     }
 }
